requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.4
djangorestframework==3.12.2
djongo==1.3.3
dnspython==1.16.0
Pillow==8.0.1
pymongo==3.7.2
pytz==2020.4
sqlparse==0.2.4
wincertstore==0.2
